pyproject.toml
[project]
name = "yt-srt-gen"
description = "Download YouTube videos, generate and translate subtitles"
version = "0.1.0"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Kirill Fertikov", email = "kirill.fertikov@gmail.com" }
]
dependencies = [
    "yt-dlp==2025.12.8",
    "openai-whisper==20250625",
    "googletrans==4.0.2",
]

[project.scripts]
yt-srt-gen = "yt_srt_gen.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
